requirements.txt
requests==2.31.0
pydantic==1.10.8
tenacity==8.2.2
asyncio==3.4.3
python-dotenv==1.0.0
logging==0.4.9.6
aiohttp>=3.9.0
aiofiles==23.2.1
Pillow>=9.5.0
SQLAlchemy>=2.0.27
dataclasses==0.6
openai>=1.12.0

# CLI dependencies
click>=8.1.7
colorama>=0.4.6  # For Windows color support

# Updated for Python 3.13 compatibility 
# Database drivers - install the one you need
# PostgreSQL
# psycopg[binary]>=3.1.18
# MySQL
# aiomysql>=0.2.0

# Testing dependencies - uncomment to install
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# pytest-mock>=3.11.1
# pytest-xdist>=3.3.1
# pytest-timeout>=2.1.0
# pytest-benchmark>=4.0.0
# freezegun>=1.2.0
# coverage>=7.3.2
# hypothesis>=6.75.3